added equirect utils
